协查通样式提交

+++ guns-vip-main/src/main/java/cn/stylefeng/guns/modular/investigation/controller/InvestigationUnitController.java
@@ -4,12 +4,18 @@ import cn.stylefeng.guns.base.pojo.page.LayuiPageInfo;
 import cn.stylefeng.guns.modular.investigation.entity.InvestigationUnit;
 import cn.stylefeng.guns.modular.investigation.model.params.InvestigationUnitParam;
 import cn.stylefeng.guns.modular.investigation.service.InvestigationUnitService;
+import cn.stylefeng.guns.sys.modular.system.entity.User;
+import cn.stylefeng.guns.sys.modular.system.model.UserDto;
+import cn.stylefeng.guns.sys.modular.system.service.UserService;
 import cn.stylefeng.roses.core.base.controller.BaseController;
 import cn.stylefeng.roses.kernel.model.response.ResponseData;
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
+
+import java.util.List;
 
 
 /**
@@ -22,6 +28,8 @@ import org.springframework.web.bind.annotation.ResponseBody;
 @RequestMapping("/investigationUnit")
 public class InvestigationUnitController extends BaseController {
 
+    @Autowired
+    UserService userService;
     private String PREFIX = "/investigationUnit";
 
     @Autowired
@@ -69,7 +77,12 @@ public class InvestigationUnitController extends BaseController {
     @RequestMapping("/addItem")
     @ResponseBody
     public ResponseData addItem(InvestigationUnitParam investigationUnitParam) {
+        List<InvestigationUnit> unit_leader_phoneNum = investigationUnitService.list(new QueryWrapper<InvestigationUnit>().eq("unit_leader_phoneNum", investigationUnitParam.getUnitLeaderPhonenum()));
+        if(unit_leader_phoneNum.size()>0){
+            return ResponseData.error("手机号码不能相同");
+        }
         this.investigationUnitService.add(investigationUnitParam);
+
         return ResponseData.success();
     }
 
@@ -95,6 +108,10 @@ public class InvestigationUnitController extends BaseController {
     @RequestMapping("/delete")
     @ResponseBody
     public ResponseData delete(InvestigationUnitParam investigationUnitParam) {
+        List<User> specialty = userService.list(new QueryWrapper<User>().eq("specialty", investigationUnitParam.getUnitId()));
+        if(specialty.size()>0){
+            return ResponseData.error("单位下有用户不可删除");
+        }
         this.investigationUnitService.delete(investigationUnitParam);
         return ResponseData.success();
     }
@@ -123,7 +140,18 @@ public class InvestigationUnitController extends BaseController {
     public LayuiPageInfo list(InvestigationUnitParam investigationUnitParam) {
         return this.investigationUnitService.findPageBySpec(investigationUnitParam);
     }
-
+    /**
+     * 查看详情接口
+     *
+     * @author hujt
+     * @Date 2020-09-09
+     */
+    @RequestMapping("/listUnit")
+    @ResponseBody
+    public ResponseData listUnit() {
+        List<InvestigationUnit> list = this.investigationUnitService.list();
+        return ResponseData.success(list);
+    }
 }
